cast

+++ easy_image.cc
@@ -291,9 +291,9 @@ void img::EasyImage::draw_zbuf_triag(ZBuffer& zbuffer, const Vector3D& A, const 
 
 	for (std::vector<lib3d::Light>::iterator it = lights.begin(); it != lights.end(); it++)
 	{
-		ambientColor.red = (uint8_t)std::min((unsigned int)255, (unsigned int)ambientColor.red + roundToInt(it->ambientLight[0] * ambientReflection[0] * 255));
-		ambientColor.green = (uint8_t)std::min((unsigned int)255, (unsigned int)ambientColor.green + roundToInt(it->ambientLight[1] * ambientReflection[1] * 255));
-		ambientColor.blue = (uint8_t)std::min((unsigned int)255, (unsigned int)ambientColor.blue + roundToInt(it->ambientLight[2] * ambientReflection[2] * 255));
+		ambientColor.red = (uint8_t)std::min((unsigned int)255, (unsigned int)ambientColor.red + (unsigned int)roundToInt(it->ambientLight[0] * ambientReflection[0] * 255));
+		ambientColor.green = (uint8_t)std::min((unsigned int)255, (unsigned int)ambientColor.green + (unsigned int)roundToInt(it->ambientLight[1] * ambientReflection[1] * 255));
+		ambientColor.blue = (uint8_t)std::min((unsigned int)255, (unsigned int)ambientColor.blue + (unsigned int)roundToInt(it->ambientLight[2] * ambientReflection[2] * 255));
 	}
 
 	for (unsigned int y_cur = y_min; y_cur <= y_max; y_cur++)
@@ -345,9 +345,9 @@ void img::EasyImage::draw_zbuf_triag(ZBuffer& zbuffer, const Vector3D& A, const 
 						double scalar_cos_alpha = (normal.x * -currentPixelVector_FromLight.x) + (normal.y * -currentPixelVector_FromLight.y) + (normal.z * -currentPixelVector_FromLight.z);
 						if (scalar_cos_alpha > 0)
 						{
-							pixelColor.red = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.red + roundToInt(it->diffuseLight[0] * diffuseReflection[0] * 255 * scalar_cos_alpha));
-							pixelColor.green = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.green + roundToInt(it->diffuseLight[1] * diffuseReflection[1] * 255 * scalar_cos_alpha));
-							pixelColor.blue = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.blue + roundToInt(it->diffuseLight[2] * diffuseReflection[2] * 255 * scalar_cos_alpha));
+							pixelColor.red = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.red + (unsigned int)roundToInt(it->diffuseLight[0] * diffuseReflection[0] * 255 * scalar_cos_alpha));
+							pixelColor.green = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.green + (unsigned int)roundToInt(it->diffuseLight[1] * diffuseReflection[1] * 255 * scalar_cos_alpha));
+							pixelColor.blue = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.blue + (unsigned int)roundToInt(it->diffuseLight[2] * diffuseReflection[2] * 255 * scalar_cos_alpha));
 						}
 
 						Vector3D r = ((2 * scalar_cos_alpha) * normal) + currentPixelVector_FromLight;
@@ -356,9 +356,9 @@ void img::EasyImage::draw_zbuf_triag(ZBuffer& zbuffer, const Vector3D& A, const 
 						double scalar_cos_beta = (r.x * -currentPixelVector_FromEye.x) + (r.y * -currentPixelVector_FromEye.y) + (r.z * -currentPixelVector_FromEye.z);
 						if (scalar_cos_beta > 0)
 						{
-							pixelColor.red = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.red + roundToInt(it->specularLight[0] * specularReflection[0] * 255 * std::pow(scalar_cos_beta, reflectionCoeff)));
-							pixelColor.green = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.green + roundToInt(it->specularLight[1] * specularReflection[1] * 255 * std::pow(scalar_cos_beta, reflectionCoeff)));
-							pixelColor.blue = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.blue + roundToInt(it->specularLight[2] * specularReflection[2] * 255 * std::pow(scalar_cos_beta, reflectionCoeff)));
+							pixelColor.red = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.red + (unsigned int)roundToInt(it->specularLight[0] * specularReflection[0] * 255 * std::pow(scalar_cos_beta, reflectionCoeff)));
+							pixelColor.green = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.green + (unsigned int)roundToInt(it->specularLight[1] * specularReflection[1] * 255 * std::pow(scalar_cos_beta, reflectionCoeff)));
+							pixelColor.blue = (uint8_t)std::min((unsigned int)255, (unsigned int)pixelColor.blue + (unsigned int)roundToInt(it->specularLight[2] * specularReflection[2] * 255 * std::pow(scalar_cos_beta, reflectionCoeff)));
 						}
 					}
 				}
